Move ISR related global variables to isr.h
